Refactor metrics.rs
